+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Issue [#522](https://github.com/m-marini/wheellyj/issues/522): Add relocate robot on stalemate
+
 ## [0.21.0] 2025-02-0
 
 ### Added